refactor: passport 코드 분리
