+++ package.json
@@ -26,6 +26,7 @@
     "express-session": "^1.18.1",
     "mongodb": "^6.13.0",
     "passport": "^0.7.0",
-    "passport-local": "^1.0.0"
+    "passport-local": "^1.0.0",
+    "socket.io": "^4.8.1"
   }
 }
